add more seeders, left: activity transactions + dealing with transactionseeder and userlimits
this is so much fun :)

+++ database/migrations/2020_02_26_184807_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name')->unique();
             $table->float('price');
-            $table->integer('category_id');
+            $table->unsignedBigInteger('category_id');
             $table->foreign('category_id')->references('id')->on('categories');
             $table->boolean('pst')->default(false);
             $table->integer('stock');
